Split TopLevelInternal test.
