+++ src/WTG.Analyzers.Test/TestData/VisibilityAnalyzer/TopLevelInternal/Compilation/Diagnostics.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<diagnostics id="WTG1006" message="Our convention is to omit the 'internal' modifier on types where it is already the default." severity="Hidden">
+	<diagnostic>
+		<location>Test0.cs: (3,1-9)</location>
+	</diagnostic>
+	<diagnostic>
+		<location>Test0.cs: (7,1-9)</location>
+	</diagnostic>
+	<diagnostic>
+		<location>Test0.cs: (11,1-9)</location>
+	</diagnostic>
+	<diagnostic>
+		<location>Test0.cs: (15,1-9)</location>
+	</diagnostic>
+	<diagnostic>
+		<location>Test0.cs: (19,1-9)</location>
+	</diagnostic>
+</diagnostics>
