+++ src/WTG.Analyzers.Test/TestData/VisibilityAnalyzer/TopLevelInternal/Compilation/Result.cs
@@ -1,0 +1,19 @@
+using System;
+
+class BobClass2
+{
+}
+
+struct BobStruct2
+{
+}
+
+interface BobInterface2
+{
+}
+
+enum BobEnum2
+{
+}
+
+delegate void BobDelegate2();
