+++ src/WTG.Analyzers.Test/TestData/VisibilityAnalyzer/TopLevelInternal/Compilation/Source.cs
@@ -1,0 +1,19 @@
+using System;
+
+internal class BobClass2
+{
+}
+
+internal struct BobStruct2
+{
+}
+
+internal interface BobInterface2
+{
+}
+
+internal enum BobEnum2
+{
+}
+
+internal delegate void BobDelegate2();
