+++ src/WTG.Analyzers.Test/TestData/VisibilityAnalyzer/TopLevelInternal/Namespace/Result.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 
 namespace BobNamespace
 {
@@ -20,21 +20,3 @@ namespace BobNamespace
 
 	delegate void BobDelegate1();
 }
-
-class BobClass2
-{
-}
-
-struct BobStruct2
-{
-}
-
-interface BobInterface2
-{
-}
-
-enum BobEnum2
-{
-}
-
-delegate void BobDelegate2();
