+++ src/WTG.Analyzers.Test/TestData/VisibilityAnalyzer/TopLevelInternal/Namespace/Source.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 
 namespace BobNamespace
 {
@@ -20,21 +20,3 @@ namespace BobNamespace
 
 	internal delegate void BobDelegate1();
 }
-
-internal class BobClass2
-{
-}
-
-internal struct BobStruct2
-{
-}
-
-internal interface BobInterface2
-{
-}
-
-internal enum BobEnum2
-{
-}
-
-internal delegate void BobDelegate2();
